requirements.txt
streamlit
youtube-transcript-api
transformers
torch
langgraph
google-generativeai
python-dotenv
duckduckgo-search
langchain-community
pytube
pytube
